TOA-13. Login files moved to login package
